app/requirements.txt
asgiref==3.8.1
cffi==1.17.1
crispy-bootstrap5==2024.2
cryptography==43.0.1
dj-database-url==2.2.0
Django==5.0.7
django-autocomplete-light==3.11.0
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-filter==24.3
django-heroku==0.3.1
django-htmx==1.18.0
django-jazzmin==3.0.0
django-vite==3.0.5
djangorestframework==3.15.2
djangorestframework-gis==1.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
packaging==24.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
whitenoise==6.7.0
